Separated the overlay and the dialog; Removed the dismiss animation;
Dialog was separated from the overlay because the overlay's opacity was making the dialog itself semi-transparent; further, the enter animation added a weird effect to the overlay, leaving a little bit of clear background at the top of the page as the overlay animated upward.

Removed the dismiss animation because it kept the container element around while the dialog was animating away, making it *look* like you could click on elements on the page, but the container was eating those clicks. This made the application and dialog feel very unresponsive when dismissed.